+++ milkman-rest/src/main/java/milkman/ui/plugin/rest/RequestBodyPostProcessor.java
@@ -1,0 +1,51 @@
+package milkman.ui.plugin.rest;
+
+import java.util.List;
+
+public interface RequestBodyPostProcessor {
+
+    List<RequestBodyPostProcessor> ALL_PROCESSORS = List.of(
+            new MultipartRequestBodyPostProcessor(),
+            new FormDataRequestBodyPostProcessor()
+    );
+
+    boolean canProcess(String contentType);
+
+    String process(String body);
+
+
+    class MultipartRequestBodyPostProcessor implements RequestBodyPostProcessor {
+
+        @Override
+        public boolean canProcess(String contentType) {
+            return contentType.contains("multipart/");
+        }
+
+        @Override
+        public String process(String body) {
+            if (!body.contains("\r\n")) {
+                return body.replace("\n", "\r\n");
+            }
+            return body;
+        }
+    }
+
+    class FormDataRequestBodyPostProcessor implements RequestBodyPostProcessor {
+
+        @Override
+        public boolean canProcess(String contentType) {
+            return contentType.contains("application/x-www-form-urlencoded");
+        }
+
+        @Override
+        public String process(String body) {
+            return body.replace("\n", "");
+        }
+    }
+
+
+    static List<RequestBodyPostProcessor> processors() {
+        return ALL_PROCESSORS;
+    }
+
+}