+++ milkman/src/main/java/milkman/ctrl/WorkspaceController.java
@@ -572,7 +572,7 @@ public class WorkspaceController {
 		var duplicate = ObjectUtils.deepClone(request);
 		duplicate.setId(UUID.randomUUID().toString());
 		duplicate.setName("Copy of " + request.getName());
-		request.setDirty(false);
+		request.setDirty(true);
 		request.setInStorage(false);
 
 		displayRequest(duplicate);