+++ changelog.md
@@ -7,6 +7,12 @@
 
 # Changelog
 
+version 5.1.0 -- upcoming
+  * added result environment to test plugin
+  * fixed bugs in test-plugin that lead to re-execution of tests
+  * added support for line-breaks in `x-www-form-urlencoded` mimetype requests
+  * 
+
 version 5.0.0
   * included [cassandra plugin](/milkman-cassandra) and [Grpc plugin](/milkman-grpc) into macos app bundle
   * reordered toolbar-actions